+++ src/typing_ex/typing.cljs
@@ -9,8 +9,7 @@
    [reagent.dom :as rdom]
    [typing-ex.plot :refer [bar-chart]]))
 
-
-(def ^:private version "v2.8.893")
+(def ^:private version "v2.9.907")
 
 (def ^:private timeout 60)
 (def ^:private todays-limit 10)
@@ -88,12 +87,25 @@ of yonder warehouses will not suffice."])
              (= all (+ goods bads)) (+ score seconds (- bs))
              :else (- score bs)))))
 
+(defn- exam-point!
+  "check mode, "
+  [login count pt]
+  (go (let [stat (-> (<! (http/get "/stat")) :body)]
+        (when (= stat "exam")
+          (let [ret (<! (http/post
+                         "/exam"
+                         {:form-params
+                          {:__anti-forgery-token (csrf-token),
+                           :login login
+                           :count count
+                           :pt pt}}))]
+            (.log js/console (str "exam-point! /exam" ret)))))))
+
 (defn show-score
   [pt]
   (if (empty? (:results @app-state))
     (js/alert (str "コピペはダメよ"))
-    (let [;; pt (:pt @app-state)
-          login (get-login)
+    (let [login (get-login)
           s1 (str login " さんのスコアは " pt " 点です。")
           s2 (condp <= pt
                100 "すばらしい。最高点取れた？平均で 80 点越えよう。"
@@ -111,12 +123,13 @@ of yonder warehouses will not suffice."])
                    (apply str (:results @app-state))
                    "\n\n"
                    (:text  @app-state))))))
-  ;; VScode, 2024-04-26
-  ;; (when (< (:todays-trials @app-state) 3)
-  ;;   (js/alert "VScode?"))
+  ;; /alert で取れる情報(文字列)をアラートに出す。
   (go (when-let [{:keys [body]} (<! (http/get "/alert"))]
         (when (re-find #"\S" body)
           (js/alert body))))
+  ;; 試験成績を記録するならここ。
+  ;; pt @mt-counter login
+  (exam-point! (get-login) @mt-counter pt)
   ;;
   (swap! app-state update :todays-trials inc)
   (when (< todays-limit (:todays-trials @app-state))